Fix on Video, certificate and quiz panel pages codes
Okay, the fix that we did is to update the video page that was redirecting like a loop to the login page because it didn't have the time to think if the user is logging in or not. And same thing for the certificate page and same thing for the quiz panel and for the quiz panel also. We had to change some checkbox or things like that to ensure that the system can automatically recognize when the user is signing in.